+++ protobuf/lib/protobuf.dart
@@ -4,7 +4,7 @@
 
 library protobuf;
 
-import 'dart:async' show Future;
+import 'dart:async' show Future, Stream;
 import 'dart:collection' show ListBase, MapBase;
 import 'dart:convert'
     show base64Decode, base64Encode, jsonEncode, jsonDecode, Utf8Codec;